# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CJ Vehicle Tracker API",
        "version": "1"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1"
        },
        {
            "url": "http://127.0.0.1:8000/api/v1"
        },
        {
            "url": "http://127.0.0.1:8000/api/v1"
        },
        {
            "url": "http://127.0.0.1:8000/api/v1"
        },
        {
            "url": "http://127.0.0.1:8000/api/v1"
        },
        {
            "url": "http://127.0.0.1:8000/api/v1"
        },
        {
            "url": "http://127.0.0.1:8000/api/v1"
        }
    ],
    "paths": {
...
        "/vehicles": {
            "get": {
                "tags": [
                    "Vehicles"
                ],
                "summary": "Get vehicles",
                "operationId": "d17171521b86a0b324bfa4d4f6ca1f8a",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/VehicleResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Vehicles not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Vehicles"
                ],
                "summary": "Create a new vehicle",
                "operationId": "e19336f3d4ff20d9858a84d5113213bc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreVehicleRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Vehicle created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/VehicleResource"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/vehicles/get-paginated": {
            "get": {
                "tags": [
                    "Vehicles"
                ],
                "summary": "Get paginated vehicles",
                "operationId": "035a1a8caef2fd1ec9c640a247b3237f",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": 20
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": 1
                        }
                    },
                    {
                        "name": "vehicleNo",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "wpac-1212"
                        }
                    },
                    {
                        "name": "chassisNo",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "123-abc-456"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/VehiclePaginatedCollection"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Vehicles not found"
                    }
                }
            }
        },
        "/vehicles/{id}": {
            "get": {
                "tags": [
                    "Vehicles"
                ],
                "summary": "Get vehicle",
                "operationId": "10887c4dacfad2ebc95acc8bcf1b7188",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the vehicle to retrieve",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/VehicleResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Vehicle not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Vehicles"
                ],
                "summary": "Update a vehicle",
                "operationId": "5a154eaaf029bc3a95ff0cac19946fc9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the vehicle to update",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateVehicleRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Vehicle updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/VehicleResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Vehicle not found"
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Vehicles"
                ],
                "summary": "Delete a vehicle",
                "operationId": "6c00c88c53d96dd1f08e69ba6d6484a1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the vehicle to delete",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Vehicle deleted successfully"
                    },
                    "404": {
                        "description": "Vehicle not found"
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/vehicles/{id}/is-deletable": {
            "get": {
                "tags": [
                    "Vehicles"
                ],
                "summary": "Check vehicle delability",
                "operationId": "bed188b75f3fa836d190219e89265c83",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the vehicle",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DeletabilityResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Vehicle not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreVehicleRequest": {
                "title": "Vehicle Create Request",
                "required": [
                    "mileageTrackerId",
                    "vehicleNo",
                    "chassisNo",
                    "currentMileage"
                ],
                "properties": {
                    "mileageTrackerId": {
                        "type": "string",
                        "example": "1"
                    },
                    "vehicleNo": {
                        "type": "string",
                        "example": "WPAC3434"
                    },
                    "chassisNo": {
                        "type": "string",
                        "example": "7788-334433"
                    },
                    "currentMileage": {
                        "type": "interger",
                        "example": "300"
                    }
                },
                "type": "object"
            },
            "UpdateVehicleRequest": {
                "title": "Vehicle Update Request",
                "required": [
                    "mileageTrackerId",
                    "vehicleNo",
                    "chassisNo",
                    "currentMileage"
                ],
                "properties": {
                    "mileageTrackerId": {
                        "type": "string",
                        "example": "1"
                    },
                    "vehicleNo": {
                        "type": "string",
                        "example": "WPAC3434"
                    },
                    "chassisNo": {
                        "type": "string",
                        "example": "7788-334433"
                    },
                    "currentMileage": {
                        "type": "interger",
                        "example": "300"
                    }
                },
                "type": "object"
            },
...
            "DeletabilityResource": {
                "title": "Deletability Resource",
                "properties": {
                    "isDeletable": {
                        "type": "boolean",
                        "example": "false"
                    },
                    "messages": {
                        "example": "[]"
                    }
                },
                "type": "object"
            },
...
            "VehiclePaginatedCollection": {
                "title": "Vehicle Paginated Collection",
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/VehicleResource"
                        }
                    },
                    "total": {
                        "type": "integer",
                        "format": "int64",
                        "example": 20
                    }
                },
                "type": "object"
            },
            "VehicleResource": {
                "title": "Vehicle Resource",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": "1"
                    },
                    "mileageTrackerId": {
                        "type": "string",
                        "example": "1"
                    },
                    "vehicleNo": {
                        "type": "string",
                        "example": "WPAC3434"
                    },
                    "chassisNo": {
                        "type": "string",
                        "example": "7788-334433"
                    },
                    "currentMileage": {
                        "type": "interger",
                        "example": "300"
                    }
                },
                "type": "object"
            }
        }
    },
...
}